+++ lib/src/types/types.dart
@@ -1,7 +1,6 @@
 
 /// Describe the equality of two item
 typedef Equals<T> = bool Function(T it1, T it2);
-bool defaultEquals<T>(T it1, T it2) => it1 == it2;
 
 /// Describe how state update when event come
 typedef Reduce<State, Event> = State Function(State state, Event event);